+++ src/Symfony/Bundle/TwigBundle/DependencyInjection/Compiler/ExceptionListenerPass.php
@@ -15,7 +15,7 @@ use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
 /**
- * Registers the Twig exception listener if Twig is registered as a templating engine.
+ * Registers the Twig exception listener if required dependencies do exist.
  *
  * @author Fabien Potencier <fabien@symfony.com>
  */
@@ -27,14 +27,9 @@ class ExceptionListenerPass implements CompilerPassInterface
             return;
         }
 
-        // register the exception controller only if Twig is enabled and required dependencies do exist
+        // register the exception controller only if required dependencies do exist
         if (!class_exists('Symfony\Component\Debug\Exception\FlattenException') || !interface_exists('Symfony\Component\EventDispatcher\EventSubscriberInterface')) {
             $container->removeDefinition('twig.exception_listener');
-        } elseif ($container->hasParameter('templating.engines')) {
-            $engines = $container->getParameter('templating.engines');
-            if (!\in_array('twig', $engines)) {
-                $container->removeDefinition('twig.exception_listener');
-            }
         }
     }
 }