+++ src/Symfony/Bundle/TwigBundle/CHANGELOG.md
@@ -6,6 +6,8 @@ CHANGELOG
 
  * updated default value for the `strict_variables` option to `%kernel.debug%` parameter
  * removed support to load templates from the legacy directories `src/Resources/views/` and `src/Resources/<BundleName>/views/`
+ * removed `TwigEngine` class, use `Twig\Environment` instead
+ * removed `FilesystemLoader` and `NativeFilesystemLoader`, use Twig notation for templates instead
 
 4.2.0
 -----